requirements.txt
Flask==3.1.0 
tensorflow==2.19.0
numpy==2.2.4
pandas==2.2.3 
Pillow==11.2.1 
kagglehub==0.3.11
tqdm==4.67.1